requirements.txt
apns2==0.5.0
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-cors-headers==3.2.0
django-extensions==2.2.5
django-heroku==0.3.1
django-push-notifications==1.6.1
django-rest-auth==0.9.5
django-sass-processor==0.7.5
djangorestframework==3.11.0
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
py-vapid==1.7.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.10.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
